Move db and json_db into app dir
