+++ app/json_db/db.js
@@ -1,6 +1,6 @@
 import fs from'fs';
 
-import { logTool } from '../app/utils/index.js';
+import { logTool } from '../utils/index.js';
